Revert forced table merge to one-card layout; drop redundant Provider List page
The previous change collapsed OAuth and API key credentials into one
table with shared columns, which lost the fine-tuned per-type column
layout (issuer/expiry vs. base URL/masked key) for no real benefit —
the actual ask was one card, not one table. Restored OAuthTable and
ApiKeyTable as-is and put both back under a single Surface, separated
by a Divider with compact inline empty states, so the page reads as
one "Credentials" card while each table keeps its tuned columns.

Also removed the standalone Provider List page/route and its private
components (ProviderTable, ProviderFilterBar, ProtocolBadge) — it only
displayed a read-only, non-actionable provider catalog, and Onboarding
already renders a fuller browsable catalog (search + connect inline).
The Credentials page's "Providers" button now links to /onboarding
instead of the removed /credentials/providers route.
